+++ kolibri/core/assets/src/api-resources/contentNodeProgress.js
@@ -1,0 +1,17 @@
+const Resource = require('../api-resource').Resource;
+
+class ContentNodeProgressResource extends Resource {
+  static resourceName() {
+    return 'contentnodeprogress';
+  }
+  static idKey() {
+    return 'pk';
+  }
+  static resourceIdentifiers() {
+    return [
+      'channel_id',
+    ];
+  }
+}
+
+module.exports = ContentNodeProgressResource;