+++ kolibri/plugins/management/assets/src/views/manage-content-page/index.vue
@@ -49,10 +49,7 @@
         </table>
       </div>
     </template>
-    <template v-else>
-      {{ $tr('notAdmin') }}
-    </template>
-
+    <auth-message v-else :header="$tr('notAdminHeader')" :details="$tr('notAdminDetails')" />
 
   </div>
 
@@ -73,9 +70,11 @@
       import: 'Import',
       export: 'Export',
       noChannels: 'No channels installed',
-      notAdmin: 'You need to sign in as the Device Owner to manage content. (This is the account originally created in the Setup Wizard.)',
+      notAdminHeader: 'You need to sign in as the Device Owner to manage content',
+      notAdminDetails: 'The Device Owner is the account originally created in the Setup Wizard',
     },
     components: {
+      'auth-message': require('kolibri.coreVue.components.authMessage'),
       'icon-button': require('kolibri.coreVue.components.iconButton'),
       'task-status': require('./task-status'),
       'wizard-import-source': require('./wizard-import-source'),