+++ CHANGELOG.rst
@@ -7,7 +7,10 @@ Changes are ordered reverse-chronologically.
 
 0.5
 ---
+
  - Update all user logging related timestamps to a custom datetime field that includes timezone info
+ - Added daemon mode (system service) to run ``kolibri start`` in background (default!) #1548
+ - Implemented ``kolibri stop`` and ``kolibri status`` #1548
 
 
 0.4