+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ Release Notes
 
 Changes are ordered reverse-chronologically.
 
+0.5
+---
+ - Update all user logging related timestamps to a custom datetime field that includes timezone info
+ - Newly imported channels are given a 'last_updated' timestamp
+ - Add progress annotation for topics, lazily loaded to increase page load performance
+
 
 0.5
 ___